pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autonomous-analytics-agent"
version = "0.1.0"
description = "Multi-agent pipeline that takes a dataset + business question and autonomously produces a professional HTML report."
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "duckdb>=1.1.0",
    "plotly>=5.24.0",
    "pandas>=2.2.0",
    "pydantic>=2.9.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-cov",
    "pre-commit",
    "pandas-stubs",
    "types-jinja2",
]

[project.scripts]
analytics-agent = "analytics_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/analytics_agent"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.mypy]
strict = true
python_version = "3.11"
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests (require API key)",
]
